The Unstable-API part of the rasterization test code
The bulk of this diff depends on enhancements submitted for FreeType
( the "diagnostics patch" ) and three enhancements to SharpFont
( https://github.com/Robmaister/SharpFont/pull/84 ). Most of the rest
is some by-pass code that talks to FreeType directly without going through
SharpFont; this should eventually ends up in SharpFont. As such, essentially
the whole of this diff would be rewritten and revised, and should not be
depended on.

cp -p bin/SharpFont.dll.config bin/Compat.dll.config; in preparation of DllImport